API (Exported Functions)

Dedisp.dedisp!

Dedisp.dedisp!

Dedisp.plan_chunked_dedisp

Dedisp.plan_dedisp

Dedisp.standardize

Dedisp.Δt

Dedisp.Δt

Dedisp.dedisp! — Method

dedisp!(output,source,plan)

Dedisperses dynamic spectra source in-place into output according to the dedispersion plan plan.

source has dimensions time samples x frequency channels

output has dimensions starting time samples x DM trials

plan has dimensions frequency channels x DM trials

Dedisp.dedisp! — Method

dedisp!(output, source, chunked_plan)

Identical to dedisp! except operates on chunked plans. Returned is the reduced output.

Dedisp.plan_chunked_dedisp — Method

plan_chunked_dedisp(freqs, f_max, dms, δt, n_chunk)

(GPU-Only) Identical to plan_dedisp except chunks the frequencies by n_chunk.

Dedisp.plan_dedisp — Method

plan_dedisp(freqs, f_max, dms, δt)

Create a dedispersion plan for that covers pulses dispersed over freqs with a maximum of f_max in MHz over all possible DMs in dms with source time resolution δt

Dedisp.standardize — Method

standardize(A)

Transform matrix A into units of signal to noise ratio along optional dimension dim, defaulting to the first dimension.

Dedisp.Δt — Method

Δt(f_min, f_max, DM, δt)

Given two frequency bounds, f_min and f_max in MHz, get the time shift in samples for a dispersed pulse of dispersion measure DM.

Dedisp.Δt — Method

Δt(f_min, f_max, DM)

Given two frequency bounds, f_min and f_max in MHz, get the time shift in seconds for a dispersed pulse of dispersion measure DM.